refact: Move tests into their own class
