+++ tests/Blueprint/BlueprintTest.php
@@ -39,6 +39,15 @@ class BlueprintTest extends TestCase
 		Dir::remove(static::TMP);
 	}
 
+	public function testAcceptRules(): void
+	{
+		$blueprint = new Blueprint([
+			'model' => $this->model,
+		]);
+
+		$this->assertInstanceOf(AcceptRules::class, $blueprint->acceptRules());
+	}
+
 	public function testConstructWithoutModel(): void
 	{
 		$this->expectException(InvalidArgumentException::class);
@@ -108,216 +117,6 @@ class BlueprintTest extends TestCase
 
 		$this->assertSame($expected, $blueprint->toArray()['tabs']);
 		$this->assertSame($expected['main'], $blueprint->tab());
-	}
-
-	public function testAcceptedFileTemplatesDefault(): void
-	{
-		$blueprint = new Blueprint([
-			'model' => $this->model,
-			'name'  => 'default',
-			'sections' => [
-				'files' => [
-					'type' => 'files',
-				],
-			]
-		]);
-
-		$this->assertSame(['default'], $blueprint->acceptedFileTemplates());
-	}
-
-	public function testAcceptedFileTemplatesFromFields(): void
-	{
-		$this->app = new App([
-			'roots' => [
-				'index' => '/dev/null'
-			],
-			'blueprints' => [
-				'files/a' => [
-					'name' => 'a',
-				],
-				'files/b' => [
-					'name' => 'b',
-				],
-				'files/c' => [
-					'name' => 'c',
-				],
-				'files/d' => [
-					'name' => 'd',
-				],
-				'files/e' => [
-					'name' => 'e',
-				],
-			]
-		]);
-
-		$blueprint = new Blueprint([
-			'model' => $this->model,
-			'name'  => 'default',
-			'fields' => [
-				'a' => [
-					'type' => 'files',
-					'uploads' => [
-						'template' => 'a'
-					]
-				],
-				'b' => [
-					'type' => 'textarea',
-					'uploads' => [
-						'template' => 'b'
-					]
-				],
-				'c' => [
-					'type'   => 'structure',
-					'fields' => [
-						'text' => [
-							'type' => 'textarea',
-							'uploads' => [
-								'template' => 'c'
-							]
-						]
-					]
-				],
-				'd' => [
-					'type'   => 'object',
-					'fields' => [
-						'text' => [
-							'type' => 'textarea',
-							'uploads' => [
-								'template' => 'd'
-							]
-						]
-					]
-				],
-				'e' => [
-					'type' => 'blocks',
-					'fieldsets' => [
-						'text' => [
-							'fields' => [
-								'text' => [
-									'type' => 'textarea',
-									'uploads' => [
-										'template' => 'e'
-									]
-								]
-							]
-						]
-					]
-				]
-			]
-		]);
-
-		$this->assertSame(['a', 'b', 'c', 'd', 'e'], $blueprint->acceptedFileTemplates());
-	}
-
-	public function testAcceptedFileTemplatesFromFieldsAndSections(): void
-	{
-		$this->app = new App([
-			'roots' => [
-				'index' => '/dev/null'
-			],
-			'blueprints' => [
-				'files/a' => [
-					'name' => 'a',
-				],
-				'files/b' => [
-					'name' => 'b',
-				],
-				'files/c' => [
-					'name' => 'c',
-				],
-			]
-		]);
-
-		$blueprint = new Blueprint([
-			'model' => $this->model,
-			'name'  => 'default',
-			'sections' => [
-				'fields' => [
-					'type' => 'fields',
-					'fields' => [
-						'a' => [
-							'type' => 'files',
-							'uploads' => [
-								'template' => 'a'
-							]
-						],
-						'b' => [
-							'type' => 'textarea',
-							'uploads' => [
-								'template' => 'b'
-							]
-						],
-					],
-				],
-				'files' => [
-					'type'     => 'files',
-					'template' => 'c'
-				]
-			]
-		]);
-
-		$this->assertSame(['a', 'b', 'c'], $blueprint->acceptedFileTemplates());
-	}
-
-	public function testAcceptedFileTemplatesFromSection(): void
-	{
-		$this->app = new App([
-			'roots' => [
-				'index' => '/dev/null'
-			],
-			'blueprints' => [
-				'files/a' => [
-					'name' => 'a',
-				]
-			]
-		]);
-
-		$blueprint = new Blueprint([
-			'model' => $this->model,
-			'name'  => 'default',
-			'sections' => [
-				'a' => [
-					'type'     => 'files',
-					'template' => 'a'
-				],
-			]
-		]);
-
-		$this->assertSame(['a'], $blueprint->acceptedFileTemplates('a'));
-	}
-
-	public function testAcceptedFileTemplatesFromSections(): void
-	{
-		$this->app = new App([
-			'roots' => [
-				'index' => '/dev/null'
-			],
-			'blueprints' => [
-				'files/a' => [
-					'name' => 'a',
-				],
-				'files/b' => [
-					'name' => 'b',
-				],
-			]
-		]);
-
-		$blueprint = new Blueprint([
-			'model' => $this->model,
-			'name'  => 'default',
-			'sections' => [
-				'a' => [
-					'type'     => 'files',
-					'template' => 'a'
-				],
-				'b' => [
-					'type'     => 'files',
-					'template' => 'b'
-				]
-			]
-		]);
-
-		$this->assertSame(['a', 'b'], $blueprint->acceptedFileTemplates());
 	}
 
 	public function testButtons(): void
